finalizando testing

+++ src/app.py
@@ -71,11 +71,13 @@ def get_planets():
     try:
         results_query = Planet.query.all()
         results = list(map(lambda item: item.serialize(), results_query))
-        response_body = {
+        if results_query:
+            response_body = {
             "msg": "Ok",
-            "results": results
-        }
-        return jsonify(response_body), 200
+            "results": results_query.serialize()
+            }
+            return jsonify(response_body), 200
+        return jsonify({"msg": "No hay planetas"}), 404
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
 
@@ -84,11 +86,13 @@ def get_planets():
 def get_planet(id):
     try:
         results_query = Planet.query.filter_by(id=id).first()
-        response_body = {
+        if results_query:
+            response_body = {
             "msg": "Ok",
             "results": results_query.serialize()
-        }
-        return jsonify(response_body), 200
+            }
+            return jsonify(response_body), 200
+        return jsonify({"msg": "No existe ese planeta"}), 404
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
 
@@ -97,8 +101,10 @@ def create_user():
     try:
         data = request.get_json()
         existing_user = User.query.filter_by(email=data["email"]).first()
+        
         if existing_user:
             return jsonify({"msg": "Email already exists"}), 400
+
         user_created = User(name=data["name"], email=data["email"], password=data["password"], is_active=data["is_active"])
         db.session.add(user_created)
         db.session.commit()
@@ -119,20 +125,25 @@ def get_characters():
             "msg": "Ok",
             "results": results
         }
-        return jsonify(response_body), 200
+        if results_query:
+            return jsonify(response_body), 200
+        return jsonify({"msg": "No existen personajes"}), 404
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
 
+        
 #GET CHARACTER BY ID
 @app.route('/character/<int:id>', methods=['GET'])
 def get_character(id):
     try:
         results_query = Character.query.filter_by(id=id).first()
-        response_body = {
-            "msg": "Ok",
-            "results": results_query.serialize()
-        }
-        return jsonify(response_body), 200
+        if results_query: 
+            response_body = {
+                "msg": "Ok",
+                "results": results_query.serialize()
+            }
+            return jsonify(response_body), 200
+        return jsonify({"msg": "Personaje no encontrado"}), 404
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
 
@@ -156,6 +167,12 @@ def get_favorites(id):
 @app.route('/users/<int:user_id>/favorites/planet/<int:planet_id>', methods=['POST'])
 def create_favorite_planet(user_id, planet_id):
     try:
+        user_exists = User.query.filter_by(id=user_id).first()
+        if not user_exists:
+            return jsonify({"msg": "El usuario no existe"}), 400
+        planet_exists = Planet.query.filter_by(id=planet_id).first()
+        if not planet_exists:
+            return jsonify({"msg": "El planeta no existe"}), 400
         existing_favorite = Favorite.query.filter_by(user_id=user_id, planet_id=planet_id).first()
         if existing_favorite:
             return jsonify({"msg": "El favorito ya existe"}), 400
@@ -174,6 +191,12 @@ def create_favorite_planet(user_id, planet_id):
 @app.route('/users/<int:user_id>/favorites/character/<int:character_id>', methods=['POST'])
 def create_favorite_character(user_id, character_id):
     try:
+        user_exists = User.query.filter_by(id=user_id).first()
+        if not user_exists:
+            return jsonify({"msg": "El usuario no existe"}), 400
+        character_exists = Character.query.filter_by(id=character_id).first()
+        if not character_exists:
+            return jsonify({"msg": "El character no existe"}), 400
         existing_favorite = Favorite.query.filter_by(user_id=user_id, character_id=character_id).first()
         if existing_favorite:
             return jsonify({"msg": "El favorito ya existe"}), 400
@@ -192,12 +215,18 @@ def create_favorite_character(user_id, character_id):
 @app.route('/users/<int:user_id>/favorites/planet/<int:planet_id>', methods=['DELETE'])
 def delete_favorite_planet(user_id, planet_id):
     try:
+        user_exists = User.query.filter_by(id=user_id).first()
+        if not user_exists:
+            return jsonify({"msg": "El usuario no existe"}), 400
+        planet_exists = Planet.query.filter_by(id=planet_id).first()
+        if not planet_exists:
+            return jsonify({"msg": "El planeta no existe"}), 400
         existing_favorite = Favorite.query.filter_by(user_id=user_id, planet_id=planet_id).first()
         if existing_favorite:
             db.session.delete(existing_favorite)
             db.session.commit()
             return jsonify({"msg": "El favorito fue eliminado"}), 200
-        return jsonify({"msg": "No se pudo eliminar"}), 400
+        return jsonify({"msg": "No existe ese favorito, no se pudo eliminar"}), 400
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
 
@@ -205,12 +234,18 @@ def delete_favorite_planet(user_id, planet_id):
 @app.route('/users/<int:user_id>/favorites/character/<int:character_id>', methods=['DELETE'])
 def delete_favorite_character(user_id, character_id):
     try:
+        user_exists = User.query.filter_by(id=user_id).first()
+        if not user_exists:
+            return jsonify({"msg": "El usuario no existe"}), 400
+        character_exists = Character.query.filter_by(id=character_id).first()
+        if not character_exists:
+            return jsonify({"msg": "El character no existe"}), 400
         existing_favorite = Favorite.query.filter_by(user_id=user_id, character_id=character_id).first()
         if existing_favorite:
             db.session.delete(existing_favorite)
             db.session.commit()
             return jsonify({"msg": "El favorito fue eliminado"}), 200
-        return jsonify({"msg": "No se pudo eliminar"}), 400
+        return jsonify({"msg": "Ese favorito no existe, no se pudo eliminar"}), 400
     except Exception as e:
         return jsonify({"msg": str(e)}), 500
